Update SineMax.py
After the recent change in the code there is no need to have a limit on the iteration. Just that there is one limit on f(x) which is that on the bit length of 7 the maximum achievable value of f(x) is  0.9999331159794518. So if you increase the value above that the loop will become infinite for l = 7. If you want more precision increase the value of l.